# grapevine-client/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "0.1.0",
    "title": "grapevine API",
    "description": "Decentralized Data Feeds Platform API"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/v1/wallets/{wallet_id}": {
      "get": {
        "tags": [
          "Wallets"
        ],
        "summary": "Get wallet by ID",
        "description": "Retrieve detailed information about a specific wallet by its unique identifier",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "wallet_id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Wallet details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "created_at": {
                      "type": "integer",
                      "minimum": 0,
                      "exclusiveMinimum": true
                    },
                    "updated_at": {
                      "type": "integer",
                      "minimum": 0,
                      "exclusiveMinimum": true
                    },
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "wallet_address": {
                      "type": "string",
                      "minLength": 42,
                      "maxLength": 42,
                      "pattern": "^0x[0-9a-fA-F]{40}$"
                    },
                    "wallet_address_network": {
                      "type": "string",
                      "enum": [
                        "base",
                        "base-sepolia",
                        "ethereum",
                        "ethereum-sepolia",
                        "polygon",
                        "polygon-amoy"
                      ]
                    },
                    "username": {
                      "type": "string",
                      "nullable": true,
                      "minLength": 3,
                      "maxLength": 50
                    },
                    "picture_url": {
                      "type": "string",
                      "nullable": true,
                      "maxLength": 2048,
                      "format": "uri"
                    }
                  },
                  "required": [
                    "created_at",
                    "updated_at",
                    "id",
                    "wallet_address",
                    "wallet_address_network",
                    "username",
                    "picture_url"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Wallet not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 100
                    },
                    "message": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 500
                    },
                    "details": {
                      "nullable": true
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 100
                    },
                    "message": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 500
                    },
                    "details": {
                      "nullable": true
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ]
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Wallets"
        ],
        "summary": "Update wallet",
        "description": "Update wallet properties including username and picture_url. Users can only update their own wallet. Requires wallet authentication via x-auth headers.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "wallet_id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 42,
              "maxLength": 42,
              "pattern": "^0x[0-9a-fA-F]{40}$",
              "description": "Ethereum wallet address (0x prefixed)",
              "example": "0x742d35Cc6634C0532925a3b844Bc9e7595f0bEb"
            },
            "required": true,
            "description": "Ethereum wallet address (0x prefixed)",
            "name": "x-wallet-address",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "pattern": "^0x[0-9a-fA-F]+$",
              "description": "Cryptographic signature (hex format)",
              "example": "0x..."
            },
            "required": true,
            "description": "Cryptographic signature (hex format)",
            "name": "x-signature",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "The signed message (base64 encoded if it contains newlines)",
              "example": "R3JhcGV2aW5lIEF1dGhlbnRpY2F0aW9u..."
            },
            "required": true,
            "description": "The signed message (base64 encoded if it contains newlines)",
            "name": "x-message",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Unix timestamp in seconds",
              "example": "1735689600"
            },
            "required": true,
            "description": "Unix timestamp in seconds",
            "name": "x-timestamp",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Chain ID for network detection (optional). Supported: 8453 (base), 84532 (base-sepolia), 1 (ethereum), 11155111 (ethereum-sepolia), 137 (polygon), 80002 (polygon-amoy)",
              "example": "8453"
            },
            "required": false,
            "description": "Chain ID for network detection (optional). Supported: 8453 (base), 84532 (base-sepolia), 1 (ethereum), 11155111 (ethereum-sepolia), 137 (polygon), 80002 (polygon-amoy)",
            "name": "x-chain-id",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string",
                    "maxLength": 50
                  },
                  "picture_url": {
                    "type": "string",
                    "format": "uri"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Wallet updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "created_at": {
                      "type": "integer",
                      "minimum": 0,
                      "exclusiveMinimum": true
                    },
                    "updated_at": {
                      "type": "integer",
                      "minimum": 0,
                      "exclusiveMinimum": true
                    },
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "wallet_address": {
                      "type": "string",
                      "minLength": 42,
                      "maxLength": 42,
                      "pattern": "^0x[0-9a-fA-F]{40}$"
                    },
                    "wallet_address_network": {
                      "type": "string",
                      "enum": [
                        "base",
                        "base-sepolia",
                        "ethereum",
                        "ethereum-sepolia",
                        "polygon",
                        "polygon-amoy"
                      ]
                    },
                    "username": {
                      "type": "string",
                      "nullable": true,
                      "minLength": 3,
                      "maxLength": 50
                    },
                    "picture_url": {
                      "type": "string",
                      "nullable": true,
                      "maxLength": 2048,
                      "format": "uri"
                    }
                  },
                  "required": [
                    "created_at",
                    "updated_at",
                    "id",
                    "wallet_address",
                    "wallet_address_network",
                    "username",
                    "picture_url"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - signature verification failed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 100
                    },
                    "message": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 500
                    },
                    "details": {
                      "nullable": true
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - you can only update your own wallet",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 100
                    },
                    "message": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 500
                    },
                    "details": {
                      "nullable": true
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Wallet not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 100
                    },
                    "message": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 500
                    },
                    "details": {
                      "nullable": true
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 100
                    },
                    "message": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 500
                    },
                    "details": {
                      "nullable": true
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/v1/wallets/address/{address}": {
      "get": {
        "tags": [
          "Wallets"
        ],
        "summary": "Get wallet by address",
        "description": "Retrieve detailed information about a specific wallet by its Ethereum wallet address (0x-prefixed, 42 characters)",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 42,
              "maxLength": 42
            },
            "required": true,
            "name": "address",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Wallet details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "created_at": {
                      "type": "integer",
                      "minimum": 0,
                      "exclusiveMinimum": true
                    },
                    "updated_at": {
                      "type": "integer",
                      "minimum": 0,
                      "exclusiveMinimum": true
                    },
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "wallet_address": {
                      "type": "string",
                      "minLength": 42,
                      "maxLength": 42,
                      "pattern": "^0x[0-9a-fA-F]{40}$"
                    },
                    "wallet_address_network": {
                      "type": "string",
                      "enum": [
                        "base",
                        "base-sepolia",
                        "ethereum",
                        "ethereum-sepolia",
                        "polygon",
                        "polygon-amoy"
                      ]
                    },
                    "username": {
                      "type": "string",
                      "nullable": true,
                      "minLength": 3,
                      "maxLength": 50
                    },
                    "picture_url": {
                      "type": "string",
                      "nullable": true,
                      "maxLength": 2048,
                      "format": "uri"
                    }
                  },
                  "required": [
                    "created_at",
                    "updated_at",
                    "id",
                    "wallet_address",
                    "wallet_address_network",
                    "username",
                    "picture_url"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Wallet not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 100
                    },
                    "message": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 500
                    },
                    "details": {
                      "nullable": true
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 100
                    },
                    "message": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 500
                    },
                    "details": {
                      "nullable": true
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  }
}